Put CAS Urls in application.yml and split controller in interface and class for logic
